+++ bfv/params.go
@@ -4,24 +4,11 @@ import (
 	"errors"
 	"fmt"
 	"math"
-	"math/bits"
 
 	"github.com/ldsec/lattigo/v2/ring"
 	"github.com/ldsec/lattigo/v2/rlwe"
 	"github.com/ldsec/lattigo/v2/utils"
 )
-
-// MaxLogN is the log2 of the largest supported polynomial modulus degree.
-const MaxLogN = 16
-
-// MinLogN is the log2 of the smallest supported polynomial modulus degree (needed to ensure the NTT correctness).
-const MinLogN = 4
-
-// MaxModuliCount is the largest supported number of moduli in the RNS representation.
-const MaxModuliCount = 34
-
-// MaxModuliSize is the largest bit-length supported for the moduli in the RNS representation.
-const MaxModuliSize = 60
 
 const (
 	// PN12QP109 is a set of parameters with N = 2^12 and log(QP) = 109
@@ -43,13 +30,9 @@ const (
 	PN15QP827pq
 )
 
-// DefaultSigma is the default error distribution standard deviation
-const DefaultSigma = 3.2
-
 type Parameters interface {
 	rlwe.Parameters
-	Moduli() *Moduli
-	LogModuli() *LogModuli
+
 	T() uint64
 	RingT() *ring.Ring
 }
@@ -62,7 +45,7 @@ var DefaultParams = []*ParametersDef{
 		t:     65537,
 		qi:    []uint64{0x7ffffec001, 0x8000016001}, // 39 + 39 bits
 		pi:    []uint64{0x40002001},                 // 30 bits
-		sigma: DefaultSigma,
+		sigma: rlwe.DefaultSigma,
 	},
 
 	{
@@ -70,7 +53,7 @@ var DefaultParams = []*ParametersDef{
 		t:     65537,
 		qi:    []uint64{0x3fffffffef8001, 0x4000000011c001, 0x40000000120001}, // 54 + 54 + 54 bits
 		pi:    []uint64{0x7ffffffffb4001},                                     // 55 bits
-		sigma: DefaultSigma,
+		sigma: rlwe.DefaultSigma,
 	},
 
 	{
@@ -79,7 +62,7 @@ var DefaultParams = []*ParametersDef{
 		qi: []uint64{0x100000000060001, 0x80000000068001, 0x80000000080001,
 			0x3fffffffef8001, 0x40000000120001, 0x3fffffffeb8001}, // 56 + 55 + 55 + 54 + 54 + 54 bits
 		pi:    []uint64{0x80000000130001, 0x7fffffffe90001}, // 55 + 55 bits
-		sigma: DefaultSigma,
+		sigma: rlwe.DefaultSigma,
 	},
 
 	{
@@ -90,7 +73,7 @@ var DefaultParams = []*ParametersDef{
 			0x3ffffffffc10001, 0x3ffffffffbe0001, 0x3ffffffffbd0001, // 58 + 58 + 58 bits
 			0x4000000004d0001, 0x400000000570001, 0x400000000660001}, // 58 + 58 + 58 bits
 		pi:    []uint64{0xffffffffffc0001, 0x10000000001d0001, 0x10000000006e0001}, // 60 + 60 + 60 bits
-		sigma: DefaultSigma,
+		sigma: rlwe.DefaultSigma,
 	},
 
 	{ // LogQP = 101.00005709794536
@@ -98,7 +81,7 @@ var DefaultParams = []*ParametersDef{
 		t:     65537,
 		qi:    []uint64{0x800004001, 0x800008001}, // 2*35
 		pi:    []uint64{0x80014001},               // 1*31
-		sigma: DefaultSigma,
+		sigma: rlwe.DefaultSigma,
 	},
 
 	{ // LogQP = 201.99999999994753
@@ -106,7 +89,7 @@ var DefaultParams = []*ParametersDef{
 		t:     65537,
 		qi:    []uint64{0x7fffffffe0001, 0x7fffffffcc001, 0x3ffffffffc001}, // 2*51 + 50
 		pi:    []uint64{0x4000000024001},                                   // 50
-		sigma: DefaultSigma,
+		sigma: rlwe.DefaultSigma,
 	},
 
 	{ // LogQP = 410.9999999999886
@@ -114,7 +97,7 @@ var DefaultParams = []*ParametersDef{
 		t:     65537,
 		qi:    []uint64{0x7fffffffff18001, 0x8000000000f8001, 0x7ffffffffeb8001, 0x800000000158001, 0x7ffffffffe70001}, // 5*59
 		pi:    []uint64{0x7ffffffffe10001, 0x400000000068001},                                                          // 59+58
-		sigma: DefaultSigma,
+		sigma: rlwe.DefaultSigma,
 	},
 
 	{ // LogQP = 826.9999999999509
@@ -124,44 +107,8 @@ var DefaultParams = []*ParametersDef{
 			0x7ffffffffb00001, 0x800000000890001, 0x8000000009d0001, 0x7ffffffff630001, 0x800000000a70001,
 			0x7ffffffff510001}, // 11*59
 		pi:    []uint64{0x800000000b80001, 0x800000000bb0001, 0xffffffffffc0001}, // 2*59+60
-		sigma: DefaultSigma,
+		sigma: rlwe.DefaultSigma,
 	},
-}
-
-// Moduli stores the NTT primes of the RNS representation.
-type Moduli struct {
-	Qi []uint64 // Ciphertext prime moduli
-	Pi []uint64 // Keys additional prime moduli
-}
-
-// Copy creates a copy of the target Moduli.
-func (m *Moduli) Copy() Moduli {
-
-	qi := make([]uint64, len(m.Qi))
-	copy(qi, m.Qi)
-
-	pi := make([]uint64, len(m.Pi))
-	copy(pi, m.Pi)
-
-	return Moduli{qi, pi}
-}
-
-// LogModuli stores the bit-length of the NTT primes of the RNS representation.
-type LogModuli struct {
-	LogQi []uint64 // Ciphertext prime moduli bit-size
-	LogPi []uint64 // Keys additional prime moduli bit-size
-}
-
-// Copy creates a copy of the target LogModuli.
-func (m *LogModuli) Copy() LogModuli {
-
-	LogQi := make([]uint64, len(m.LogQi))
-	copy(LogQi, m.LogQi)
-
-	LogPi := make([]uint64, len(m.LogPi))
-	copy(LogPi, m.LogPi)
-
-	return LogModuli{LogQi, LogPi}
 }
 
 type parameters struct {
@@ -179,7 +126,7 @@ type ParametersDef struct {
 }
 
 func NewParametersFromParamDef(paramDef *ParametersDef) (*parameters, error) {
-	m := new(Moduli)
+	m := new(rlwe.Moduli)
 	m.Qi = make([]uint64, len(paramDef.qi))
 	copy(m.Qi, paramDef.qi)
 	m.Pi = make([]uint64, len(paramDef.pi))
@@ -188,20 +135,20 @@ func NewParametersFromParamDef(paramDef *ParametersDef) (*parameters, error) {
 }
 
 // NewParametersFromModuli creates a new Parameters struct and returns a pointer to it.
-func NewParametersFromModuli(logN int, m *Moduli, t uint64) (p *parameters, err error) {
+func NewParametersFromModuli(logN int, m *rlwe.Moduli, t uint64) (p *parameters, err error) {
 
 	p = new(parameters)
 
-	if logN < MinLogN || logN > MaxLogN {
+	if logN < rlwe.MinLogN || logN > rlwe.MaxLogN {
 		return nil, fmt.Errorf("invalid polynomial ring log degree: %d", logN)
 	}
 
 	// Checks if Moduli is valid
-	if err = checkModuli(m, logN); err != nil {
+	if err = rlwe.CheckModuli(m, logN); err != nil {
 		return nil, err
 	}
 
-	p.ParametersStruct = *rlwe.NewRLWEParameters(logN, m.Qi, m.Pi, DefaultSigma)
+	p.ParametersStruct = *rlwe.NewRLWEParameters(logN, m.Qi, m.Pi, rlwe.DefaultSigma)
 
 	p.t = t
 
@@ -209,277 +156,20 @@ func NewParametersFromModuli(logN int, m *Moduli, t uint64) (p *parameters, err 
 }
 
 // NewParametersFromLogModuli creates a new Parameters struct and returns a pointer to it.
-func NewParametersFromLogModuli(logN int, lm *LogModuli, t uint64) (p *parameters, err error) {
+func NewParametersFromLogModuli(logN int, lm *rlwe.LogModuli, t uint64) (p *parameters, err error) {
 
-	if err = checkLogModuli(lm); err != nil {
+	if err = rlwe.CheckLogModuli(lm); err != nil {
 		return nil, err
 	}
 
 	// If LogModuli is valid and then generates the moduli
-	return NewParametersFromModuli(logN, genModuli(lm, logN), t)
+	return NewParametersFromModuli(logN, rlwe.GenModuli(lm, logN), t)
 }
-
-// // LogN returns the log of the degree of the polynomial ring
-// func (p *ParametersDef) LogN() uint64 {
-// 	return p.logN
-// }
-
-// // N returns power of two degree of the ring
-// func (p *ParametersDef) N() uint64 {
-// 	return 1 << p.logN
-// }
 
 // T returns the plaintext coefficient modulus t
 func (p *parameters) T() uint64 {
 	return p.t
 }
-
-// // Sigma returns standard deviation of the noise distribution
-// func (p *ParametersDef) Sigma() float64 {
-// 	return p.sigma
-// }
-
-// // SetT sets the plaintext coefficient modulus t
-// func (p *ParametersDef) SetT(T uint64) {
-// 	p.t = T
-// }
-
-// // WithT returns a copy of the parmaters with a plaintext modulus set to T
-// func (p *ParametersDef) WithT(T uint64) (pCopy *ParametersDef) {
-// 	pCopy = p.Copy()
-// 	pCopy.SetT(T)
-// 	return
-// }
-
-// LogModuli generates a LogModuli struct from the parameters' Moduli struct and returns it.
-func (p *parameters) LogModuli() (lm *LogModuli) {
-	lm = new(LogModuli)
-	lm.LogQi = make([]uint64, len(p.Q()))
-	for i := range p.Q() {
-		lm.LogQi[i] = uint64(math.Round(math.Log2(float64(p.Q()[i]))))
-	}
-	lm.LogPi = make([]uint64, len(p.P()))
-	for i := range p.P() {
-		lm.LogPi[i] = uint64(math.Round(math.Log2(float64(p.P()[i]))))
-	}
-	return
-}
-
-// Moduli returns a struct Moduli with the moduli of the parameters
-func (p *parameters) Moduli() (m *Moduli) {
-	m = new(Moduli)
-	m.Qi = make([]uint64, len(p.Q()))
-	copy(m.Qi, p.Q())
-	m.Pi = make([]uint64, len(p.P()))
-	copy(m.Pi, p.P())
-	return
-}
-
-// // Q returns a new slice with the factors of the ciphertext modulus q
-// func (p *ParametersDef) Q() []uint64 {
-// 	qi := make([]uint64, len(p.qi))
-// 	copy(qi, p.qi)
-// 	return qi
-// }
-
-// // QCount returns the number of factors of the ciphertext modulus q
-// func (p *ParametersDef) QCount() uint64 {
-// 	return uint64(len(p.qi))
-// }
-
-// func (p *ParametersDef) QBigInt() *big.Int {
-// 	q := big.NewInt(1)
-// 	for _, qi := range p.qi {
-// 		q.Mul(q, new(big.Int).SetUint64(qi))
-// 	}
-// 	return q
-// }
-
-// // P returns a new slice with the factors of the ciphertext modulus extension P
-// func (p *ParametersDef) P() []uint64 {
-// 	pi := make([]uint64, len(p.pi))
-// 	copy(pi, p.pi)
-// 	return pi
-// }
-
-// // PCount returns the number of factors of the ciphertext modulus extension P
-// func (p *ParametersDef) PCount() uint64 {
-// 	return uint64(len(p.pi))
-// }
-
-// func (p *ParametersDef) PBigInt() *big.Int {
-// 	pInt := big.NewInt(1)
-// 	for _, pi := range p.pi {
-// 		pInt.Mul(pInt, new(big.Int).SetUint64(pi))
-// 	}
-// 	return pInt
-// }
-
-// func (p *ParametersDef) QP() []uint64 {
-// 	qp := make([]uint64, len(p.qi)+len(p.pi))
-// 	copy(qp, p.qi)
-// 	copy(qp[len(p.qi):], p.pi)
-// 	return qp
-// }
-
-// // QPCount returns the number of factors of the ciphertext modulus Q + the modulus extension P
-// func (p *ParametersDef) QPCount() uint64 {
-// 	return p.QCount() + p.PCount()
-// }
-
-// func (p *ParametersDef) QPBigInt() *big.Int {
-// 	pqInt := p.QBigInt()
-// 	pqInt.Mul(pqInt, p.PBigInt())
-// 	return pqInt
-// }
-
-// // LogQP returns the size of the extended modulus QP in bits
-// func (p *ParametersDef) LogQP() uint64 {
-// 	tmp := ring.NewUint(1)
-// 	for _, qi := range p.qi {
-// 		tmp.Mul(tmp, ring.NewUint(qi))
-// 	}
-// 	for _, pi := range p.pi {
-// 		tmp.Mul(tmp, ring.NewUint(pi))
-// 	}
-// 	return uint64(tmp.BitLen())
-// }
-
-// // LogQ returns the size of the modulus Q in bits
-// func (p *ParametersDef) LogQ() uint64 {
-// 	tmp := ring.NewUint(1)
-// 	for _, qi := range p.qi {
-// 		tmp.Mul(tmp, ring.NewUint(qi))
-// 	}
-// 	return uint64(tmp.BitLen())
-// }
-
-// // LogP returns the size of the modulus P in bits
-// func (p *ParametersDef) LogP() uint64 {
-// 	tmp := ring.NewUint(1)
-// 	for _, pi := range p.pi {
-// 		tmp.Mul(tmp, ring.NewUint(pi))
-// 	}
-// 	return uint64(tmp.BitLen())
-// }
-
-// // LogQAlpha returns the size in bits of the sum of the norm of
-// // each element of the special RNS decomposition basis for the
-// // key-switching.
-// // LogQAlpha is the size of the element that is multiplied by the
-// // error during the keyswitching and then divided by P.
-// // LogQAlpha should be smaller than P or the error added during
-// // the key-switching wont be negligible.
-// func (p *ParametersDef) LogQAlpha() uint64 {
-
-// 	alpha := p.PCount()
-
-// 	if alpha == 0 {
-// 		return 0
-// 	}
-
-// 	res := ring.NewUint(0)
-// 	var j uint64
-// 	for i := uint64(0); i < p.QCount(); i = i + alpha {
-
-// 		j = i + alpha
-// 		if j > p.QCount() {
-// 			j = p.QCount()
-// 		}
-
-// 		tmp := ring.NewUint(1)
-// 		for _, qi := range p.qi[i:j] {
-// 			tmp.Mul(tmp, ring.NewUint(qi))
-// 		}
-
-// 		res.Add(res, tmp)
-// 	}
-
-// 	return uint64(res.BitLen())
-// }
-
-// // Alpha returns the number of moduli in in P
-// func (p *ParametersDef) Alpha() uint64 {
-// 	return p.PCount()
-// }
-
-// // Beta returns the number of element in the RNS decomposition basis: Ceil(lenQi / lenPi)
-// func (p *ParametersDef) Beta() uint64 {
-// 	if p.Alpha() != 0 {
-// 		return uint64(math.Ceil(float64(p.QCount()) / float64(p.Alpha())))
-// 	}
-
-// 	return 0
-// }
-
-// // NewPolyQ returns a new empty polynomial of degree 2^logN in basis qi.
-// func (p *ParametersDef) NewPolyQ() *ring.Poly {
-// 	return ring.NewPoly(p.N(), p.QCount())
-// }
-
-// // NewPolyP returns a new empty polynomial of degree 2^logN in basis Pi.
-// func (p *ParametersDef) NewPolyP() *ring.Poly {
-// 	return ring.NewPoly(p.N(), p.PCount())
-// }
-
-// // NewPolyQP returns a new empty polynomial of degree 2^logN in basis qi + Pi.
-// func (p *ParametersDef) NewPolyQP() *ring.Poly {
-// 	return ring.NewPoly(p.N(), p.QPCount())
-// }
-
-// // GaloisElementForColumnRotationBy returns the galois element for plaintext
-// // column rotations by k position to the left. Providing a negative k is
-// // equivalent to a right rotation.
-// func (p *ParametersDef) GaloisElementForColumnRotationBy(k int) uint64 {
-// 	twoN := 1 << (p.logN + 1)
-// 	mask := twoN - 1
-// 	kRed := uint64(k & mask)
-// 	return ring.ModExp(GaloisGen, kRed, uint64(twoN))
-// }
-
-// // GaloisElementForRowRotation returns the galois element for generating the row
-// // rotation automorphism
-// func (p *ParametersDef) GaloisElementForRowRotation() uint64 {
-// 	return (1 << (p.logN + 1)) - 1
-// }
-
-// // GaloisElementsForRowInnerSum returns a list of all galois elements required to
-// // perform an InnerSum operation. This corresponds to all the left rotations by
-// // k-positions where k is a power of two and the row-rotation element.
-// func (p *ParametersDef) GaloisElementsForRowInnerSum() (galEls []uint64) {
-// 	galEls = make([]uint64, p.logN+1, p.logN+1)
-// 	galEls[0] = p.GaloisElementForRowRotation()
-// 	for i := 0; i < int(p.logN)-1; i++ {
-// 		galEls[i+1] = p.GaloisElementForColumnRotationBy(1 << i)
-// 	}
-// 	return galEls
-// }
-
-// // InverseGaloisElement takes a galois element and returns the galois element
-// //  corresponding to the inverse automorphism
-// func (p *ParametersDef) InverseGaloisElement(galEl uint64) uint64 {
-// 	twoN := uint64(1 << (p.logN + 1))
-// 	return ring.ModExp(galEl, twoN-1, twoN)
-// }
-
-// func (p *ParametersDef) RingQ() *ring.Ring {
-// 	ringQ, err := ring.NewRing(p.N(), p.qi)
-// 	if err != nil {
-// 		panic(err) // Parameter type invariant
-// 	}
-// 	return ringQ
-// }
-
-// func (p *ParametersDef) RingP() *ring.Ring {
-// 	if len(p.pi) == 0 {
-// 		return nil
-// 	}
-// 	ringP, err := ring.NewRing(p.N(), p.pi)
-// 	if err != nil {
-// 		panic(err) // Parameter type invariant
-// 	}
-// 	return ringP
-// }
 
 func (p *parameters) RingT() *ring.Ring {
 	ringQP, err := ring.NewRing(p.N(), []uint64{p.t})
@@ -488,20 +178,6 @@ func (p *parameters) RingT() *ring.Ring {
 	}
 	return ringQP
 }
-
-// // Copy creates a copy of the target Parameters.
-// func (p *ParametersDef) Copy() (paramsCopy *ParametersDef) {
-
-// 	paramsCopy = new(ParametersDef)
-// 	paramsCopy.logN = p.logN
-// 	paramsCopy.t = p.t
-// 	paramsCopy.sigma = p.sigma
-// 	paramsCopy.qi = make([]uint64, len(p.qi))
-// 	copy(paramsCopy.qi, p.qi)
-// 	paramsCopy.pi = make([]uint64, len(p.pi))
-// 	copy(paramsCopy.pi, p.pi)
-// 	return
-// }
 
 // Equals compares two sets of parameters for equality.
 func (p *parameters) Equals(other Parameters) (res bool) {
@@ -554,8 +230,8 @@ func (p *parameters) UnmarshalBinary(data []byte) error {
 
 	logN := int(b.ReadUint8())
 
-	if logN > MaxLogN {
-		return fmt.Errorf("logN larger than %d", MaxLogN)
+	if logN > rlwe.MaxLogN {
+		return fmt.Errorf("logN larger than %d", rlwe.MaxLogN)
 	}
 
 	lenQi := b.ReadUint8()
@@ -569,7 +245,7 @@ func (p *parameters) UnmarshalBinary(data []byte) error {
 	b.ReadUint64Slice(qi)
 	b.ReadUint64Slice(pi)
 
-	err := checkModuli(p.Moduli(), logN)
+	err := rlwe.CheckModuli(p.Moduli(), logN)
 
 	p.ParametersStruct = *rlwe.NewRLWEParameters(logN, qi, pi, sigma)
 
@@ -578,116 +254,3 @@ func (p *parameters) UnmarshalBinary(data []byte) error {
 	}
 	return nil
 }
-
-func checkModuli(m *Moduli, logN int) (err error) {
-
-	if len(m.Qi) > MaxModuliCount {
-		return fmt.Errorf("#qi is larger than %d", MaxModuliCount)
-	}
-
-	if len(m.Pi) > MaxModuliCount {
-		return fmt.Errorf("#Pi is larger than %d", MaxModuliCount)
-	}
-
-	for i, qi := range m.Qi {
-		if bits.Len64(qi)-1 > MaxModuliSize+1 {
-			return fmt.Errorf("qi bit-size for i=%d is larger than %d", i, MaxModuliSize)
-		}
-	}
-
-	for i, pi := range m.Pi {
-		if bits.Len64(pi)-1 > MaxModuliSize+1 {
-			return fmt.Errorf("Pi bit-size for i=%d is larger than %d", i, MaxModuliSize)
-		}
-	}
-
-	N := 1 << logN
-
-	for i, qi := range m.Qi {
-		if !ring.IsPrime(qi) || qi&uint64((N<<1)-1) != 1 {
-			return fmt.Errorf("qi n°%d is not an NTT prime", i)
-		}
-	}
-
-	for i, pi := range m.Pi {
-		if !ring.IsPrime(pi) || pi&uint64((N<<1)-1) != 1 {
-			return fmt.Errorf("Pi n°%d is not an NTT prime", i)
-		}
-	}
-
-	return nil
-}
-
-func checkLogModuli(lm *LogModuli) (err error) {
-
-	// Checks if the parameters are empty
-	if lm.LogQi == nil || len(lm.LogQi) == 0 {
-		return fmt.Errorf("nil or empty slice provided as LogModuli.LogQi")
-	}
-
-	if len(lm.LogQi) > MaxModuliCount {
-		return fmt.Errorf("#LogQi is larger than %d", MaxModuliCount)
-	}
-
-	if len(lm.LogPi) > MaxModuliCount {
-		return fmt.Errorf("#LogPi is larger than %d", MaxModuliCount)
-	}
-
-	for i, qi := range lm.LogQi {
-		if qi > MaxModuliSize {
-			return fmt.Errorf("LogQi for i=%d is larger than %d", i, MaxModuliSize)
-		}
-	}
-
-	for i, pi := range lm.LogPi {
-		if pi > MaxModuliSize {
-			return fmt.Errorf("LogPi for i=%d is larger than %d", i, MaxModuliSize)
-		}
-	}
-
-	return nil
-}
-
-// GenModuli generates the appropriate primes from the parameters using generateNTTPrimes such that all primes are different.
-func genModuli(lm *LogModuli, logN int) (m *Moduli) {
-
-	m = new(Moduli)
-
-	// Extracts all the different primes bit-size and maps their number
-	primesbitlen := make(map[uint64]int)
-
-	for _, qi := range lm.LogQi {
-		primesbitlen[qi]++
-	}
-
-	for _, pj := range lm.LogPi {
-		primesbitlen[pj]++
-	}
-
-	// For each bit-size, it finds that many primes
-	primes := make(map[uint64][]uint64)
-	for key, value := range primesbitlen {
-		primes[key] = ring.GenerateNTTPrimes(int(key), 2<<logN, value)
-	}
-
-	// Assigns the primes to the BFV moduli chain
-	m.Qi = make([]uint64, len(lm.LogQi))
-	for i, qi := range lm.LogQi {
-		m.Qi[i] = primes[qi][0]
-		primes[qi] = primes[qi][1:]
-	}
-
-	// Assigns the primes to the special primes list for the extended ring
-	m.Pi = make([]uint64, len(lm.LogPi))
-	for i, pj := range lm.LogPi {
-		m.Pi[i] = primes[pj][0]
-		primes[pj] = primes[pj][1:]
-	}
-
-	return
-}
-
-// // Temporary getter, rlwe.Parameters should be embeded in type bfv.parameters
-// func (p *Parameters) RLWEParameters() *rlwe.ParametersStruct {
-// 	return rlwe.NewRLWEParameters(p.logN, p.qi, p.pi, p.sigma)
-// }
